Push 3 - up to lab:05
I have been attempting to be able to push directly from intellij but I have been having a lot of trouble with it. I am now attempting to try it with github Desktop instead and hopefully this will prove to be more of  a success. In this push, I am bringing my project up to date in accordance to lab 05: Koltin Recap. I have also began to improve my model to match more of my own video format. I still have to correct some errors which I hope to correct this evening but may have to complete tomorrow instead depending on time restraints tonight and this evening.