Delete + Filtering (tags) + other improvements
Deleting function has been added along with some filtering.  The filtering is done by searching though the tags of each of the videos added. If the video includes the tag searched for, the program will print out this video. Improvements include the inclusion of more Try - Catch uses in order to stop the program from crashing with incorrect data inputs